backend: refactor auth Controller handlers util
