+++ backend/example.config.env
@@ -20,3 +20,6 @@ SUPPORT_EMAIL=""
 # JWT
 ACCESS_TOKEN_SECRET=""
 REFRESH_TOKEN_SECRET=""
+
+# frontend domain
+FRONTEND_DOMAIN=""
